CPP Module 00: 2nd try [complete] [100%]

+++ CPP Modules/00/ex02/test.sh
@@ -1,8 +1,15 @@
-clang++ -I./tests/ *.cpp ./tests/$1.cpp -o test
+file=./tests/tests_$1.cpp
 
-cat ./tests/1992*.log | cut -f2- -d] > should_output.txt
-./test | cut -f2- -d] > user_output.txt
-
-diff should_output.txt user_output.txt
-
-rm -f should_output.txt user_output.txt
+if test -f "$file"; then
+	clang++ -Werror -Wall -Wextra -std=c++98 -I./tests/ *.cpp $file -o test
+	
+	cat ./tests/1992*.log | cut -f2- -d] > should_output.txt
+	./test | cut -f2- -d] > user_output.txt
+	
+	diff should_output.txt user_output.txt
+	
+	rm -f should_output.txt user_output.txt
+else
+	echo File \'$file\' does not exists.
+	exit 1
+fi
